Remove fake leaderboard generation functionality
Previously, the performance calculator GUI, whenever faced with a
beatmap that didn't have any scores, would just make up scores on the
spot, by generating hit statistics and computing everything from that.

With the game-side switch to ScoreV2, this is no longer feasible. In
particular, in ScoreV2, the _order_ of the incoming judgements matters.
Which means that in order to generate a proper score for the fake
statistics data, the fake score has to be generated _judgement by
judgement_.

This functionality is way too finicky and complicated to bring back,
and the end result is likely to be useless. So just trim it entirely
and display a placeholder instead.